+++ src/Lifetime/__init__.py
@@ -28,8 +28,8 @@ def shutdown(exit_code,fast = 0):
     global _shutdown_timeout
     if _shutdown_phase == 0:
         # Thread safety? proably no need to care
-        import ZServer
-        ZServer.exit_code = exit_code
+        from Zope2.Startup import config
+        config.ZSERVER_EXIT_CODE = exit_code
         _shutdown_phase = 1
     if fast:
         # Someone wants us to shutdown fast. This is hooked into SIGTERM - so