+++ src/ZServer/__init__.py
@@ -11,25 +11,17 @@
 #
 ##############################################################################
 
-import sys
 import utils
 
-#########################################################
-### declarations used by external packages
-
-# the exit code used to exit a Zope process cleanly
-exit_code = 0
+from Zope2.Startup.config import (  # NOQA
+    ZSERVER_CONNECTION_LIMIT as CONNECTION_LIMIT,
+    ZSERVER_EXIT_CODE as exit_code,
+    ZSERVER_LARGE_FILE_THRESHOLD as LARGE_FILE_THRESHOLD,
+    setNumberOfThreads,
+)
 
 # the ZServer version number
 ZSERVER_VERSION = '1.1'
-
-# the maximum number of incoming connections to ZServer
-CONNECTION_LIMIT = 1000 # may be reset by max_listen_sockets handler in Zope
-
-# request bigger than this size get saved into a
-# temporary file instead of being read completely into memory
-LARGE_FILE_THRESHOLD = 1 << 19 # may be reset by large_file_threshold
-                               # handler in Zope
 
 # the Zope version string
 ZOPE_VERSION = utils.getZopeVersion()
@@ -42,11 +34,7 @@ from HTTPServer import zhttp_server, zhttp_handler
 from PCGIServer import PCGIServer
 from FCGIServer import FCGIServer
 from FTPServer import FTPServer
-from PubCore import setNumberOfThreads
 from medusa.monitor import secure_monitor_server
-
-### end declarations
-##########################################################
 
 # we need to patch asyncore's dispatcher class with a new
 # log_info method so we see medusa messages in the zLOG log
@@ -54,5 +42,3 @@ utils.patchAsyncoreLogger()
 
 # we need to patch the 'service name' of the medusa syslog logger
 utils.patchSyslogServiceName()
-
-