+++ src/Zope2/Startup/handlers.py
@@ -1,99 +1,103 @@
 import os
+import re
 import sys
-from re import compile
 from socket import gethostbyaddr
 
-# top-level key handlers
+from Zope2.Startup import config
 
-
-# XXX  I suspect there are still problems here.  In most cases, if the
-# value is not set from the config file, or if the value is the
-# default, the corresponding envvar is not unset, which is needed to
-# ensure that the environment variables and the configuration values
-# reflect a coherant configuration.
 
 def _setenv(name, value):
     if isinstance(value, str):
         os.environ[name] = value
     else:
-        os.environ[name] = `value`
+        os.environ[name] = repr(value)
+
 
 def debug_mode(value):
     value and _setenv('Z_DEBUG_MODE', '1')
     return value
+
 
 def locale(value):
     import locale
     locale.setlocale(locale.LC_ALL, value)
     return value
 
+
 def datetime_format(value):
     value and _setenv('DATETIME_FORMAT', value)
     return value
 
+
 def automatically_quote_dtml_request_data(value):
     not value and _setenv('ZOPE_DTML_REQUEST_AUTOQUOTE', '0')
     return value
+
 
 def maximum_number_of_session_objects(value):
     default = 1000
     value not in (None, default) and _setenv('ZSESSION_OBJECT_LIMIT', value)
     return value
 
+
 def session_add_notify_script_path(value):
     value is not None and _setenv('ZSESSION_ADD_NOTIFY', value)
     return value
 
+
 def session_delete_notify_script_path(value):
     value is not None and _setenv('ZSESSION_DEL_NOTIFY', value)
     return value
+
 
 def session_timeout_minutes(value):
     default = 20
     value not in (None, default) and _setenv('ZSESSION_TIMEOUT_MINS', value)
     return value
 
+
 def large_file_threshold(value):
-    import ZServer
-    ZServer.LARGE_FILE_THRESHOLD = value
+    config.ZSERVER_LARGE_FILE_THRESHOLD = value
+
 
 def http_realm(value):
     value is not None and _setenv('Z_REALM', value)
     return value
 
+
 def max_listen_sockets(value):
-    import ZServer
-    ZServer.CONNECTION_LIMIT = value
+    config.ZSERVER_CONNECTION_LIMIT = value
+
 
 def cgi_maxlen(value):
     import cgi
     cgi.maxlen = value
 
+
 def http_header_max_length(value):
     return value
 
+
 def enable_ms_public_header(value):
-    import webdav
-    webdav.enable_ms_public_header = value
+    config.ZSERVER_ENABLE_MS_PUBLIC_HEADER = value
 
-# server handlers
 
-def root_handler(config):
+def root_handler(cfg):
     """ Mutate the configuration with defaults and perform
     fixups of values that require knowledge about configuration
     values outside of their context.
     """
 
     # Set environment variables
-    for k,v in config.environment.items():
+    for k, v in cfg.environment.items():
         os.environ[k] = v
 
     # Add directories to the pythonpath
-    instancelib = os.path.join(config.instancehome, 'lib', 'python')
-    if instancelib not in config.path:
+    instancelib = os.path.join(cfg.instancehome, 'lib', 'python')
+    if instancelib not in cfg.path:
         if os.path.isdir(instancelib):
-            config.path.append(instancelib)
-    path = config.path[:]
+            cfg.path.append(instancelib)
+    path = cfg.path[:]
     path.reverse()
     for dir in path:
         sys.path.insert(0, dir)
@@ -101,56 +105,55 @@ def root_handler(config):
     # Add any product directories not already in Products.__path__.
     # Directories are added in the order they are mentioned
 
-    instanceprod = os.path.join(config.instancehome, 'Products')
-    if instanceprod not in config.products:
+    instanceprod = os.path.join(cfg.instancehome, 'Products')
+    if instanceprod not in cfg.products:
         if os.path.isdir(instanceprod):
-            config.products.append(instanceprod)
+            cfg.products.append(instanceprod)
 
     import Products
     L = []
-    for d in config.products + Products.__path__:
+    for d in cfg.products + Products.__path__:
         if d not in L:
             L.append(d)
     Products.__path__[:] = L
 
     # if no servers are defined, create default http server and ftp server
-    if not config.servers:
-        config.servers = []
+    if not cfg.servers:
+        cfg.servers = []
 
     # prepare servers:
-    for factory in config.servers:
-        factory.prepare(config.ip_address or '',
-                        config.dns_resolver,
+    for factory in cfg.servers:
+        factory.prepare(cfg.ip_address or '',
+                        cfg.dns_resolver,
                         "Zope2",
-                        config.cgi_environment,
-                        config.port_base)
+                        cfg.cgi_environment,
+                        cfg.port_base)
 
     # set up trusted proxies
-    if config.trusted_proxies:
-        from ZPublisher import HTTPRequest
-        from ZServer.medusa import http_server
-        # DM 2004-11-24: added host name mapping (such that examples in
-        # conf file really have a chance to work
+    if cfg.trusted_proxies:
         mapped = []
-        for name in config.trusted_proxies: mapped.extend(_name2Ips(name))
-        HTTPRequest.trusted_proxies = tuple(mapped)
-        http_server.trusted_proxies = tuple(mapped)
-    
-    # set the maximum number of ConflictError retries
-    if config.max_conflict_retries:
+        for name in cfg.trusted_proxies:
+            mapped.extend(_name2Ips(name))
+        config.TRUSTED_PROXIES = tuple(mapped)
+
         from ZPublisher import HTTPRequest
-        HTTPRequest.retry_max_count = config.max_conflict_retries
+        HTTPRequest.trusted_proxies = tuple(mapped)
+
+    # set the maximum number of ConflictError retries
+    if cfg.max_conflict_retries:
+        from ZPublisher import HTTPRequest
+        HTTPRequest.retry_max_count = cfg.max_conflict_retries
 
 
-def handleConfig(config, multihandler):
+def handleConfig(cfg, multihandler):
     handlers = {}
     for name, value in globals().items():
         if not name.startswith('_'):
             handlers[name] = value
     return multihandler(handlers)
 
-# DM 2004-11-24: added
-def _name2Ips(host, isIp_=compile(r'(\d+\.){3}').match):
+
+def _name2Ips(host, isIp_=re.compile(r'(\d+\.){3}').match):
     """Map a name *host* to the sequence of its ip addresses.
 
     use *host* itself (as sequence) if it already is an ip address.