+++ src/webdav/NullResource.py
@@ -105,7 +105,7 @@ class NullResource(Persistent, Implicit, Resource):
     def PUT(self, REQUEST, RESPONSE):
         """Create a new non-collection resource.
         """
-        from ZServer import LARGE_FILE_THRESHOLD
+        from Zope2.Startup.config import ZSERVER_LARGE_FILE_THRESHOLD
 
         self.dav__init(REQUEST, RESPONSE)
 
@@ -125,8 +125,9 @@ class NullResource(Persistent, Implicit, Resource):
             raise PreconditionFailed
 
         # SDS: Only use BODY if the file size is smaller than
-        # LARGE_FILE_THRESHOLD, otherwise read LARGE_FILE_THRESHOLD
-        # bytes from the file which should be enough to trigger
+        # ZSERVER_LARGE_FILE_THRESHOLD, otherwise read
+        # ZSERVER_LARGE_FILE_THRESHOLD bytes from the file
+        # which should be enough to trigger
         # content_type detection, and possibly enough for CMF's
         # content_type_registry too.
         #
@@ -142,7 +143,8 @@ class NullResource(Persistent, Implicit, Resource):
         # REQUEST['BODYFILE'] directly and try as much as possible not
         # to read the whole file into memory.
 
-        if int(REQUEST.get('CONTENT_LENGTH') or 0) > LARGE_FILE_THRESHOLD:
+        if (int(REQUEST.get('CONTENT_LENGTH') or 0) >
+                ZSERVER_LARGE_FILE_THRESHOLD):
             file = REQUEST['BODYFILE']
             body = file.read(LARGE_FILE_THRESHOLD)
             file.seek(0)